# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ruralicos API",
    "version": "1.0.0",
    "description": "Alertas de subvenciones y ayudas rurales por WhatsApp. Generado por introspeccion con scripts/generar_openapi.js — regenerar tras cambiar rutas. Todas las rutas aceptan tambien el prefijo /v1 (alias del contrato v1)."
  },
  "servers": [
    {
      "url": "https://ruralicos-api.onrender.com/v1",
      "description": "Produccion (prefijo v1 recomendado)"
    },
    {
      "url": "https://ruralicos-api.onrender.com",
      "description": "Produccion (rutas legacy sin prefijo)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerUser": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de usuario (POST /login-phone)"
      },
      "bearerOrg": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de cooperativa (POST /partner/login)"
      },
      "bearerAdmin": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de admin (POST /admin/login)"
      },
      "cronToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cron-token",
        "description": "Token de tareas internas/cron"
      }
    }
  },
  "paths": {
...
    "/partner/branding": {
      "get": {
        "tags": [
          "partner"
        ],
        "security": [
          {
            "bearerOrg": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "partner"
        ],
        "security": [
          {
            "bearerOrg": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/partner/digests": {
      "get": {
        "tags": [
          "partner"
        ],
        "security": [
          {
            "bearerOrg": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/partner/members": {
      "get": {
        "tags": [
          "partner"
        ],
        "security": [
          {
            "bearerOrg": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "partner"
        ],
        "security": [
          {
            "bearerOrg": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/partner/members/{userId}": {
      "delete": {
        "tags": [
          "partner"
        ],
        "security": [
          {
            "bearerOrg": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "partner"
        ],
        "security": [
          {
            "bearerOrg": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/partner/overview": {
      "get": {
        "tags": [
          "partner"
        ],
        "security": [
          {
            "bearerOrg": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}